requirements.txt
streamlit==1.35.0 
openai==1.30.1 
weaviate-client==4.5.5 
sentence-transformers==2.7.0 
langgraph==0.0.34 
tenacity==8.2.3 
python-dotenv==1.0.1 
requests==2.31.0
